+++ saleor/graphql/shop/mutations/shop_settings_update.py
@@ -20,7 +20,7 @@ from ...meta.inputs import MetadataInput, MetadataInputDescription
 from ...plugins.dataloaders import get_plugin_manager_promise
 from ...site.dataloaders import get_site_promise
 from ...storefront_traffic import set_allow_storefront_traffic_cache
-from ..enums import AccountConfirmModeEnum, PasswordLoginModeEnum
+from ..enums import PasswordLoginModeEnum
 from ..types import Shop
 
 
@@ -150,15 +150,6 @@ class ShopSettingsInput(graphene.InputObjectType):
             "`productVariantMetadataUpdated`) are sent."
             + ADDED_IN_322
             + DEPRECATED_IN_3X_INPUT
-        ),
-    )
-
-    account_confirm_merge_mode = AccountConfirmModeEnum(
-        description=(
-            "Controls the method used for merging existing orders and giftcards "
-            "when password-based authentication is used. "
-            "Learn more at "
-            "https://docs.saleor.io/upgrade-guides/core/migrate-account-merging"
         ),
     )