+++ saleor/graphql/shop/types.py
@@ -55,7 +55,6 @@ from ..translations.resolvers import resolve_translation
 from ..translations.types import ShopTranslation
 from ..utils import format_permissions_for_display
 from .enums import (
-    AccountConfirmModeEnum,
     AnnouncementImportanceEnum,
     GiftCardSettingsExpiryTypeEnum,
     PasswordLoginModeEnum,
@@ -503,16 +502,6 @@ class Shop(graphene.ObjectType):
         deprecation_reason=DEFAULT_DEPRECATION_REASON,
     )
 
-    account_confirm_merge_mode = AccountConfirmModeEnum(
-        description=(
-            "Controls the method used for merging existing orders and giftcards "
-            "when password-based authentication is used. "
-            "Learn more at "
-            "https://docs.saleor.io/upgrade-guides/core/migrate-account-merging"
-        ),
-        required=True,
-    )
-
     class Meta:
         description = (
             "Represents a shop resource containing general shop data and configuration."
@@ -806,8 +795,3 @@ class Shop(graphene.ObjectType):
     @load_site_callback
     def resolve_use_legacy_update_webhook_emission(_, _info, site):
         return site.settings.use_legacy_update_webhook_emission
-
-    @staticmethod
-    @load_site_callback
-    def resolve_account_confirm_merge_mode(_, _info, site):
-        return site.settings.account_confirm_merge_mode